Convert project structure to use a gradle build
* aligns the project structure with the general structure for gradle
  based java projects
* removes unneeded files

This project needs the PHP Toolbox as a compile dependency and can only
be built after the PHP Toolbox Plugin has a release (0.5.7 was assumed).

To run this project, import the gradle build file as a project from
existing sources and execute the `runIde` gradle target.

To do the same from the cli, use `./gradlew runIde`.