Add inspection for deprecated annotations + QuickFix
This adds an extensionPoint that can provide replacements.

If a replacement is provided, it replaces the import statement.

Also contains an example extension to replace the `@Route` annotation.

Could be further expanded to provide a `Configurable` with custom
replacements or further actions such as a custom QuickFix to invoke
compacting the `@Method` annotation into the new (old) `@Route`
annotation.